Bug 708485: Workaround a SEGV seen in pdf14 clist pattern filling
This is a 'safe' fix in that it's very low cost, and it will
exactly stop the case where we crash.

It is a less than ideal fix, as it is merely avoiding a crash
rather than avoiding getting into the situation that provoked the
crash in the first place.

The issue is that when running the fuzzed file under the clist,
we get to a fill pattern operation without the required part of
the pattern tile having been setup.

The non-clist code allocates this buffer by calling
pdf14_patt_trans_image_fill from pdf14_begin_typed_image.

pdf14_patt_trans_image_fill claims in a comment to be called both
from pdf14_begin_typed_image and pdf14_clist_begin_typed_image,
but it is only ever called from the former. Looking at the
history for this file, it looks like it has only ever been called
from the one place; presumably when Michael was developing this
he intended to call it from both places, but this was changed
before the commit went in, and the comment has always been out
of date!

So, the clist code reaches pdf14_clist_begin_typed_image, and
performs various checks, culminating in:

 if (dev_proc(dev, begin_typed_image) != pdf14_clist_begin_typed_image) {

When running in the failure case, we fail this check (i.e. the
begin_typed_image device proc is indeed pdf14_clist_begin_typed_image.)

Contrast this to the equivalent check in pdf14_begin_typed_image
(the non-clist code):

 if (dev_proc(dev, begin_typed_image) != gx_default_begin_typed_image) {

Running in non-clist most, the value here is pdf14_begin_typed_image, so
we call pdf14_patt_trans_image_fill.

I don't claim to understand what exactly these checks are doing,
but the appear to be mismatching.

Modifying the clist one to trigger in the case where we are running
is something worth considering. I've tried that on the cluster
(comparing against gx_default_begin_typed_image), and noticable diffs
are given (various images fail to appear in the output).

Such a modification doesn't even solve our problem case. While the code
within the if is doing something non-entirely-disimilar to the code in
pdf14_patt_trans_image_fill it's not a complete copy. We hit a problem
whereby ctx->stack is dereferenced when ctx == NULL.

I therefore propose to go with the sticking plaster fix for now. If
we ever find a case whereby a non-fuzzed file gives a difference
between clist and non-clist cases, we can revisit.