Bug 708486: Workaround segfault in clist_fill_path()
The test case has a very large negative number as the final parameter (ty) being
passed to the "cm" operator.

For the matrix concatenation values are converted from floats to doubles for the
interim calculations to preserve accuracy, then back to floats when written back
to the ctm.

In this case, the conversion from double to float produces slightly different
values on aarch64 vs x86/x86_64 etc, which ultimately results in the "cropped"
path height in clist_fill_path() ending up orders of magnitude larger than the
height of the device.

This adds a safety net of checking the cropped height is not larger than the
device's height.

This approach was preferred as a simple, quick and low risk solution, rather than
devoting the considerable time and affort to "fixing" the rounding to handle a
situation that can never result in usable output.