Bug 708507: pdfi: Duplicate glyph names in TTF post table
The post table in the embedded TTF has duplicate glyph names.

Since we now create a pdfi dictionary to hold the post table mappings, the
earlier, and correct, instances were overwritten by the later, incorrect ones.

This changes the dictionary put operation to not allow existing key/value pairs
to be overwritten.