PDF interpreter - fix Optional Content Usage processing
Bug #708503 "Overlapping Text Issue When Converting PDF to JPEG in Ghostscript 10.04"

The description is entirely misleading. The actual problem is that the
input PDF file has a number of 'layers' (Optional Content groups). One
of those layers is called 'REMOVE' and is locked, set to off in the
Usage dictionary (but 'ON' in the optional content Properties).

Due to a couple of logic bugs in the Optional Content Usage processing
it was missing the fact that the group should not be visible.