Disable clang compiler optimisation for setjmp/longjmp
We've had a few OSS-fuzz reports involving crashes with longjmp in the
JPEG error handling code, even after memset'ing the jump buffer. These
are not reproducible with other compilers (gcc or Visual Studio) and
do not reproduce with debug builds.

Attempts to investigate suggest the jump buffer is invalid/corrupted
after longjmp, leading to attempts to use invalid memory.

The suspicion is that there is something about the clang compiler
optimisation causing a problem with setjmp or longjmp. Preventing
function inline'ing does not seem to avoid the issue. This commit uses
the 'optnone' attribute to disable compiler optimisations around all
functions which use setjmp or longjmp, when the compiler is clang.

For me this resolves the remaining problems, but as with the memset of
the jump buffer I cannot say that it is a 'fix'.