OpenJPEG stream code - prevent buffer overrun
OSS-fuzz #414383025

The fuzzed file ends up processing what appears to be a YCC image, which
only has one component, I think this is potentially possible for Indexed
colour, it is certainly possible for RGB.

The code didn't check for the number of components but simply calls our
own YCC to RGB code.

Obviously that shouldn't happen because at this point we don't have YCC
colours. THe problem occurs because the width of the image is not a
multiple of 3; because we loop over the data using an unsigned counter
it is possible to decrement below 0 and the counter simply 'wraps round'
to a very large value. We then use that to index into the image data and
that accesses invalid memory.

So two fixes; firstly don't do a YCC to RGB conversion if we don't have
3 components. Secondly, in both the 8 and 16-bit YCC to RGB routines,
check the number of bytes is a multiple of the sample size. If it is
not, then reduce it until it is. That might cause a single image sample
to be incorrectly coloured but it avoids the invalid memory access.