pdfwrite - detect use of /.notdef when creating PDF/A-2+
bug #708479 "ghostscript doesn't create a PDF/A 3b compliant file, .notdef glyph included"

The PDF/A-2 spec (but not A-1) specifies that the use of the /.notdef
glyph by text showing operators is forbidden. It does not actually state
that we should not use the equivalent GID 0 from a TrueType font which
seems like something of an oversight. Nor does it mention composite or
CIDFonts. The omission of type 3 fonts makes sense, as these tend to
have no 'semantic value' anyway.

So, if we are creating PDFA-2 or more, and the font is a type 1 or
type 2 (CFF) font, then check the glyph name and if it is './notdef'
then follow the PDFACompatibilityPolicy for a creation error.

If the Policy is 1 then we will use the existing fallback method, which
will render the glyph to a bitmap, create a type 3 font to hold it, with
a manufactured name (so not /.notdef) and use that insted of the
original. Less than ideal but it will create output with the correct
appearance and stil valid for PDF/A-2+.