pdfwrite - Don't elide clipped out text in clipping text rendering modes
Bug #708543 "missing content with Tr 7"

The problem is that the file sets a clip, then sets text in text
rendering mode 7, finally an image (pure white) is drawn over the
region. The text is entirely outside the extant clipping region, so
when the intersection of the initial clip and the clip from the text is
calculated, nothing remains; the white image is therefore not rendered.

The pdfwrite problem occurs because we note that the text is fully
clipped out, and so we elide it. This causes the clip from the text to
be lost, leaving only the original clip which does not clip out the
white image. The white image is therefore drawn and obscures the
original content.

Fixed here by simply not dropping the text if it uses any clipping text rendering mode.