Fix Coverity warnings
After commits 	1b0a04edec4e786c695fd01b98e97f1538ce1fed and
191dadf620445ecbfc99076adef9586f852bd47a Coverity has started making
some new complaints. A couple were genuine typos, most are of the
usual form; we are no chekcing something which was not previously
checked, and that has tilted the scanner to recognising more of the
same problem, or detecting problems that were previously obscured.

466572 - gxipixel.c checking NULL memory allocation after using memory.
466562 - gdevopvp.c double free, fixed by setting pointer to NULL
466563 - rinkj-screen-eb.c returning unuinitialised variable
466568 - rinkj-screen-eb.c calloc use reveresed elements and sizeof
         element, which confused Coverity.
466569 - rinkj-screen-eb.c - typo in testing allocations
466571 - actually in rinkj-config.c, could leak memory
466567 - "
466565 - ghostpdf.c typo potentially dereferences NULL pointer
466564 - dscparse.c failure to update pointer after re-allocation causes
         UAF
466567
466570 - now memory allocation is checked, exit can cause memory leak