pdfwrite - permit indirect references for /FS pdfmark
Bug #708544 "10.05.1 doesn't create /AF in Catalog automatically anymore"

See the bug report for details. Essentially we are getting a /EMBED
pdfmark with a /FS key whose value is an indrect reference rather than a
direct object (dictionary).

Its not obvious to me if this is legal using a pdfmark, but it is
possible to deal with it and if users want to add a /AF array to the
Catalog dictionary (which is optional if the output is not PDF/A-3)
then there is no way to avoid it, so this commit deals with it.