Bug 708546: Tweak how pdfwrite converts shadings to images.
Previously, the code used to work by creating a device to accumulate
data to an image. It would then render the shading. For each run of
non-self-overlapping patches, it would then fill the outline of those
patches with the image. For cases where the patch bends back on
itself, the outline can't be trusted, so a mask is used instead.

The problem in the bug springs from the fact that the filled paths
scale differently to the bitmap used to fill them. As the resolution
at which the file is viewed increases, edge pixels can appear in
the filled region that weren't filled in the image.

For shadings without a background color, this can produce nasty
white pixels on the edges of the shading.

The fix here is to tweak this behaviour for shadings that don't have
backgrounds.

We now decompose the shading twice. In the first pass, we write to
the image, but don't output it. In the second pass, we don't write
to the image, but do everything else exactly as before.

This means that by the time we start outputting paths etc, we are
guaranteed to have a 'complete' rendering of the image. In this case,
when the mesh has multiple patches that abut one another (the common
case), we will now always have rendered all the patches before we
render the image; hence the instances of white pixels should be
greatly reduced.

This means that the images we write out will typically be 'larger'
than before (because previously many of them would have had lots of
white pixels in and so would have compressed well). But the default
settings for pdfwrite at least mean that we spot the images are
identical, and common them up. So pdfwrite files will generally be
smaller, and ps2write files will generally be (marginally) bigger.