Address warnings about mktemp being insecure
Because mkromfs requires the use of some functions from the graphics lib
file system code, on Unix it builds and links gp_unifs.c which contains
gp_open_scratch_file_impl() which build time optionally uses various calls
(by preference: mkstemp64, mkstemp or mktemp) to create safe names for
temporary files.

Since mkromfs never creates temp files, gp_open_scratch_file_impl() is never
called, hence we were never concerned about it falling back to the least safe
option of mktemp, and the linker warning that comes with that.

Nevertheless, we get a small but steady trickle of people informing us about
this "problem".

This fixes it so even that code included in mkroms will call either mkstemp64 or
mkstemp, if one of those is available, in preference to mktemp.