Bug 708575: Alignment issues on WASM32
Add ARCH_ALIGN_UINT64_T_MOD

Note that since uint64_t is defined in stdint_.h which depends on
genarch.c. So we cannot just define a struct with an uint64_t member and
use OFFSET_IN.

Instead, use C11 features to determine the alignment and fall back to
the pointer width, which should work on most platforms.

Add uint64_t alignement requirement to cached_char

The code member of cached char is of type gs_glyph which is a typedef
for uint64_t.

Add uint64_t alignement requirement to struct ref_s

The dummy member of ref_value of cached char is of type uint64_t.

8-byte-align CMS memory on WASM32