pdfwrite - fix crash with type 3 fonts and OCR
Bug #708547 "Crash when performing OCR on certain PDF pages"

The problem occurred because the page was using a type 3 font, we were
ending up going through the pdfwrite 'set_cache' code when rendering,
instead of the default 'set_cache', which we should have been using.

This led to us trying to use a member of the device structure which is
normally NULL, and therefore crashing.

In pdf_text_set_cache detect the fact that we are using a type 3 font
and that we are doing the rendering stage of OCR. If that happens, use
the default gs_text_set_cache() function.

Guard the original problem code by adding a test for NULL. Add similar
checks in other places where we were just expecting the member to be
non-NULL, just in case.

Fixed a compiler warning with a cast in gdevpdfm.c while pasing.