pdfwrite - remove the /pbf idiom
No bug report for this one, the report is from a custoemr. The file uses
the Microsoft Pscript5 driver 'pbf' idiom to draw a pattern using an
image. The resulting image is too small by a factor of the image
width in the x direction and image height in the y direction, leading to
the pattern becoming a flat fill.

This is caused by the idiom recognition replacing the 'pbf' procedure.

This was originally added as part of the fix for bug #695230, but it was
added to match the 'mask_pbf' replacement which is what the report there
actually used. At that time we had no specimen for the 'pbf' procedure.

Fundamentally this doesn't work, we get away with it for the 'mask_pbf'
procedure because, essentially, the examples we have for this are very
small.

The basic problem is that the image and sizing is set up for a
particular resolution (in both cases, 600 dpi) and running at a
different resolution (pdfwrite runs at 720 dpi by default) causes odd
effects. So we recalculate the pattern to an identity matrix and step
and scale the output.

Since this doesn't work for the large(r) image in the test file, remove
the idiom for pbf. We'll keep the mask_pbf one for now, but if anyone
encounters a problem with it we may remove it in future and tell users
to alter the resolution of pdfwrite instead.