PDF interpreter - don't update text matrix for text not in BT/ET
Bug #708580 "Use of Uninitialized Memory in gs_setmatrix()"

When we encounter text which is not in a BT/ET (which is illegal) we
do an implicit BT/ET round that text. The implicit ET does a 'newpath'
which invalidates the current point leading to gs_currentpoint returning
an error code unexpectedly (because we initialise the path to 0,0 we
should never have an invalid current point here).

This code is used to update the Text matrix, which is only valid between
a BT and an ET operator. Since we are effectively wrapping the text in
BT and ET there is no point in updating the text matrix.