graphics library - initialise some variables
Bug #708581 "Use of Uninitialized Memory in gx_subpath_is_rectangular()"

This was very difficult to reproduce, because of the use of the OSS-fuzz
docker instance, it does not reproduce on a standard Linux build and we
can't seem to use valgrind with the docker instance.

Fortunately, on Chris's advice, it was possible to reproduce the problem
by using single block allocations. This additionally showed
uninitialised memory being used in gx_pattern_cache_lookup() and so that
is also addressed in this commit.