graphics library - test return code before using returned data
Bug #708582 "Use of Uninitialized Memory in gx_restrict_ICC()"

The code passes cc to the tint transform procedure, but then uses cc
before checking to see if there was an error. If there's an error then
cc isn't initialised....

Simply check the code before trying to use the data. This now matches
the logic in the Separation colour space case.