Photoshop device - limit channels to device components
Bug #708583 "Uninitialized data passed to write() through psd_write_image_data()"

This is caused by a confusion over the number of actual components being
written. Because the input is PDF we should be able to decide up front
how many spot inks (if any) are present, and set up the device correctly
before running the input. Unlike PostScript where we set up the maximum
and simply hope.

However... If the input file is sufficiently damaged we can't find the
DeviceN/Separation colour spaces, so we don't add any inks to the
output. In some circumstances, even though the colour space isn't
available from the Resources dictionary, it can still be installed by
using the object number instead of a name (in this case by a Shading).

This led to the device only creating 4 components, but the Separation
information containing spots, causing the PSD device to think it was
processing 5 components.

This commit limits the PSD device setup to the number of components in
the device structre, if they would exceed it. Obviously this can lead
to some inks potentially getting lost, but that's what happens when you
process a corrupted/invalid file.