Fix overprint simulation on RGB+spots devices
No bug report, example file supplied by customer.

The reported problem was performance bu building and running on Windows
instead of lInux threw asserts in a debug build and crashed.

This is due to the transparency compositor determining the number of
components in the device two different ways and getting different
answers. That then caused the number of planes to end up as 2, which
isn't valid.

This commit sets the current device space to RGB if the device is RGB
and supports spots, previously it always set teh colour space to
some type of CMYK.

This resolves the bug for me and with this change I cannot reproduce
the performance problem.