Fix logic for EMBED pdfmark
Bug #708616"Empty embedded file"

Logic was the wrong way round in this test, if we *don't* get any
matches from sscanf then we should abort. If we get at least one match
then we can carry on.

Checked bug #708584 and it was wrong before and correct no. I definitely
teted that, so.... baffled.