Bug 708636: Tweak buffer space calculations.
Within gs, dimensions like width, height and raster are stored as
ints. To lift this would be a major upheaval as we'd have to
start to pass size_t (or more likely ssize_t) arguments for
positions in the device interface.

However, we do need to allow for buffer spaces to be calculated as
size_t's. We thought we'd done this everywhere, but evidently not.
Fix some cases here.