Bug 708639: Fix psdrgb with overprint simulation.
The previous fix (32b94907c84) in this area (psdrgb + simulation
overprint was wrong). It left us in the position of files with at
least 1 spot color appearing to work, but files without any spot
colors giving obviously corrupt results.

The PDF interpreter was incorrectly assuming that any file where
the device had at least 4 colours was CMYK (for the purposes of
transparency group setup). This meant that files with spots were
falling through the CMYK path, and hence device overprint wasn't
actually happening. Files without spots were attempting the
CMYK overprint, but with the wrong device profile in place.

The fix here is to a) ensure that the PDFI transparency setup
spots CMYK in a better manner, and b) to remove the use of the
incorrect device profile.

Also, a note regarding our expectations of what overprint simulation
in an RGB space means. We expect an overprint simulation with
a device like psdrgb to mean "perform the rendering in a CMYK
space with overprint internally, and then, at the last minute,
convert the CMYK process colors to RGB, leaving the spot colors
intact".

Theoretically, there is another possible valid interpretation of
overprint in RGB, whereby we might try to do 'overprint' with the
process components in RGB directly; for instance, overprinting blue
(0 0 1) on top of an existing color, could mean leaving the red
and green components untouched.

This is definitely NOT what we are implementing here. If this is
ever needed in future, we'll need to find a way to distinguish the
two cases.