Bug 708633: Extend overprint_copy_planes to support 16 bits too.
When using TextAlphaBits, we end up rendering to a memory device,
and then using copy_planes to copy out a reduced version. When we
are using overprint too, that goes through overprint_copy_planes.

The existing implementation only copes with 8 bit devices. We
update it here to cope with 16bits too.