Bug 708650: Copy updated spots after simulated overprint.
If we start with a psdrgb device (for example), and then try
to use simulated overprint, we create a new overprint compositor
device. Any 'additional' spots that we find during rendering
will be added into the devn_params for this compositor device.

When the compositor device finishes, we want to offer the
original device the chance to copy this updated information
back into itself. We do this by adding a new gxdso, and call
this from the PDF14_POP_DEVICE handling.