Fix Coverity warnings
After the large commit to address integer overflows with allocations
Coverity produced a number of warnings related to the change in data
types in a few places.

A mixture of prtinf formats being incorrect for the changed data type,
overflow before promotion ot the data type, an accidentally left behind
piece of debug code and an actual error (cast of the wrong variable) in
the commercial build.

Coverity IDs 473479, 473478, 473477, 473476, 473475, 473474, 473473,
473472, 473471, 473470, 473469