ps2write - 'fix' MediaRequested feature.
Bug #708657 "	pdf2ps sets media size on every page when converting Letter document in A4 locale"

This is all tied up with several other bugs that I've identified, and
changes due to several other bugs which I have not been able to find.

On each page we request the media size required for that page, if it is
not already the correct size. This means we emit a setpagedevice.

Now if we have a fixed media size (say A4) on the printer, and we have
an input file which wants Letter, then we repeatedly request Letter
media for each page, because the media isn't what we want. We never
get the media we request, so we try again on the subsequent page, etc.

That has the effect of resetting the device, which causes Duplex to fail
to work.

We used to get around that by setting an array called MediaRequested
when we called setpagedevice to set the media size. After checking the
current media size, and before calling setpagedevice to request a size,
we checked to see if MediaRequested is present, if it is, then we check
the sizes stored in that array to see if they are the ones we were about
to request. If they are, we simply drop the request. This is because
we've already requested that size, and didn't get it, so there's no
point in asking again.

Unfortunately, over time, the 'pagesave' (the VM state at the start of
the page) has moved around and defeated the feature. Not too surprising
after 12 years.

In fact it has moved around even more since then. This commit simply
moves it again. The save state now no longer encapsulates the media
size and orientation, that is permitted to change with every page. This
shouldn't be a problem because we check the current media size against
the required size for every page. The end of job restore at the end of
the server loop on a printer or PostScript interpreter should prevent
this carrying over to subsequent jobs.

I checked the relevant bugs where I could find them, and regression
testing shows no problems with this.