Various devices - make ICC profile inclusion optional
Bug 708641 "Converting from pdf to tiff doesn't keep the CMYK"

The problem here is that we embed an ICC profile in the output of
TIFF, PNG and JPEG files, and there is no option to disable it.

The custoemr specifically uses an 'uncalibrated' workflow, they specify
exact amounts of CMYK, produce a TIFF file with those values, and then
take that CMYK for further processing. As part of that processing they
embed the TIFF into another document which has been prepared with the
same CMYK values.

But because we embed an ICC profile into the TIFF file, the workflow
uses that profile to colour manage the CMYK values. The end result is
that the TIFF colours don't match the colours from the later processing.

This commit simply allows the user to disable embedding the Artifex
output profile in the output. There were already many circumstances
under which we didn't embed the profile in the output, but there was no
simple way for a user to choose that.