PDF interpreter - improve caching for some files
No bug report, the file came from a customer. I think we have seen
files of this form before in public, but I can't find an example.

In general each page of a PDF file contains a Resources dictionary; this
dictionary contains references to all the Fonts, Images, etc which are
used by that page. Resource dictionaries are usually relatively small
since they normally contain only the Resources used by a single page.

The PDF file here has a *single* Resources dictionary which is referenced
by every page. The Resources dictionary contains a reference to every
Font, Image etc used by any page. The file in question has 2700+ pages
and so ends up with some 200,000 objects being referenced from this
single dictionary.

Previously this was a problem for us, because we check the Resources
on every page to see if the page uses transparency, spot colours etc.
There are so many objects in the Resources dictionary that we end up
evicting the dictionary from the cache while checking its contents. Then
we run the page and again we need to check the Resources, and look up
the dictionary to find the resources which are used. This leads to us
reading all 200,000 objects from the PDF file again.

We do this twice for every page, which causes very poor performance.

In this commit; alter the dict_first() and dict_next() functions so that
like the dict_get() function, if they dereference an object they store
the dereferenced object back to the dictionary, in place of the original
reference.

We cannot do this for Fonts, because Fonts are loaded lazily (ie only
when actually used) and if we were to replace the indirect references
in the Resources dictionary, the loading wouldn't take place and we
would fail to find fonts. So add a couple of new functions, used only
by the font checking code currently, dict_first_no_store_R and
dict_next_no_store_R. Like the similarly named dict_get_no_store_R and
friends these work as before and don't replace the references.

This gets resolved when (if) the fonts are actually used.

Create a new function to add a PDF object to the object cache; if it
is already in the cache promote the cache entry, otherwise add it to the
cache.

Use the above function in pdfi_check_Resources() to make sure the
current Resources dictionary is in the cache. If the Resources
dictioanry is large enough then dereferencing all its contents can end
up pushing the actual Resources dictionary out of the cache. We do this
to make sure that, at the end of the page, the Resources dictionary is
still avaiable, in case the next page uses the same dictionary. If it
does not then the cached one will eventually disappear.

Do the same at the end of pdfi_check_page(); belt and braces approach.

Extras:

Improve the error checking in pdfi_check_Resources(); if a particular
resource type (eg ColorSpace) is not present, then that's not an error.

Add a little more information to the cache debugging, if enabled.

Since we're making changes here anyway, and this problem was originally
resolved for the cutomer by altering the cache size at build time; make
the PDF maximum object cache size configurable at run-time with a
PDFCACHE control. Defaults to the original valule of 200. Document the
new control.