pdfwrite - use correct memory allocator for deferred Optional Content
Bug #708704 " use Ghostscrip t to convert the color pdf to gray pdf and get exit value 139"

The problem that I can see is that we should be using non-GC'ed memory
to store the 'pending' Optional Content string. Ordinarily this isn't
important but if a GC mark-and-sweep occurs before we use/free the
deferred optional content then it will be left pointing at invalid
memory.