Automate forcing the larger table contents format on AIX
By default, the AIX linker creates a table of contents no larger than 64Kb,
which is (and has been for some time) too small for Ghostscript.

This removes the requirement for the user to add "bigtoc" option to either the
configure or make command line options by setting it automatically in
configure.