OSS-fuzz 435802190
The changes for performance in files with ridiculously large, shared,
Resources meant that we now cache dereferenced objects when iterating
through a dictionary. That causes us to fail to spot subsequent
circular references.

Fixed here by replacing circular references in dictionaries with null
objects when iterating. Since we could not have dereferenced the
original object, we lose nothing by discarding it.