Graphics library - prevent write after free in pattern cleanup
Bug #708721 "Security: heap-use-after-free write in pattern_paint_cleanup_core during PaintProc error handling"

The code in question here was introduced in commit
9d6e533f6486fe3fd0a6e29b23d1811361adf904, ironically in response to a
bug report from OSS-fuzz.

Due to the fact that the original input was a PDF file, and we no longer
process PDF files using PostScript, it is no longer possible to trigger
the original bug.

Reverting to the nearest release didn't help, because the original
report was using memory sanitizer, and msan triggers on a different
issue before reaching the code addressed by the commit. Attempting to
work around this rapidly turned into a fruitless game of whack-a-mole.

Debugging through the code (but without msan) showed that it was
possible to reach the point of the change with a device which was still
referenced. However, the device's close method *already* set the target
pointer to NULL, so the code here seems to be redundant.

Given that, in a working case, the code results in a write-after-free
(because the device is closed and freed before reaching this point), the
original inspiration for the problem cannot happen because the code has
changed, and it is non-trivial to try and work out what purpose the
code change may have had, I've chosen to remove it after discussing it
with the original author.

If Google somehow finds a way to trigger the original problem, but
from PostScript rather than PDF, then we can reconsider.