Prevent integer overflow in LIPS device
Bug 708731 "Security: heap-buffer-overflow write due to integer overflow in lprn_print_image"

Although the report uses the npdl device, that falls through to the
more generic lips device. Ideally we'd sanity check the media size in
the initial device, but unfortunately that delegates much of its
parameter handling to lips, which is used by multiple devices, so cannot
be expected to know what is a sane size.

So check for the overflow and fail with a rangecheck if it occurs. This
of course has the benefit of working with other devices which use the
lips device and don't check their parameters.

Since we don't own the copyright in these devices that's about as far as
I'm willing to go on changing it.