ESC/P 2 device - handle totally incompressible data
Bug #708743 "Security: heap-buffer-overflow write in escp2_print_page due to insufficient buffer allocation for compression"

If the rendered data is completely incompressible, the device could
end up writing over the end of the buffer, because we actually require
an extra (size/128) bytes of RLE compressed data to represent the
original.

Obviously no real-world data would fit this category. We just allocate
an extra 'band size /128' (rounded up) bytes of buffer in case the
input data is not compressible.