Coverity CID 487677: add bounds check for error (and warning) string index
We use a byte arrays as a bitfields to keep track of what errors and warnings
have tripped while interpreting a PDF. That means there are potentially unused
bits in the final byte.

Since we use the size of that byte array to derive the loop limits as we print
the relevant errors and warnings, it can loop past the number of indices in
the lists of error and warning strings.

In practice, those extra bits should never be set and thus we'd never try to
access those non-existent error/warning strings. But that is too subtle for
Coverity to understand, and there is the potential for memory corruption to
occur.

So add a limit check and an assert, just in case.