Coverity IDs 487700, 487701, 487702
It was theoretically possible to return from setting the equivalent
ICC space for a CIE space without setting the equivalent and without
returning an error if the colour_space_index was not recognised, or the
colour space wasn't a PostScript CIE based space.

In practice I don't think this is actually possible, but this commit
alters those cases to errors. I think this should silence the warnings