pdfwrite - Improve Acrobat/Photoshop duotones
Bug #708734 "Inline image gets inverted in PDF"

The problem here is that the image is defined as a Duotone (actually
it is **not** a duotone), eg:

/PhotoshopDuotoneColorSpace [/DeviceN [(Black)][/DeviceGray] {} bind

The tint transform produces the wrong result (inverted) so if we embed
the space as is, and the device does not support /Black (eg its a
display) then the colour will be inverted.

To be clear; the fault is the PostScript program trying to be clever for
no good reason.

Because the program has an alternative definition for the colour space,
used when /Black isn't supported, and tries to determine which space to
use, we can force it to use the alternate definition by lying about the
current ColorConversionStrategy. If we say its sRGB (even if it isn't)
then the program will use the alternate definition.

Ugly hackery, it doesn't produce 'correct' output, but then the input
is incorrect before we start. This at least looks correct no matter
whether you render to RGB or CMYK.