PDF interpreter - sanitise halftone parameters
Bug #708754, 708755, 708756

We weren't checking to see that the width/height of threshold-defined
halftones were greater than 0 (where appropriate). I don't want to limit
the width and height to 32767 (as suggested in the bug report) as I feel
that's too small. The spec says these are 'integers' so using the
implementation limits in Appendix C, I've limited these to 2^32 - 1

We also weren't handling an integer overflow calculating the size of a
type 16 halftone, fix that too by promoting the intermediate variables
to 64-bit integers.

We still can't have a length > 4GB (due to the data type in use) but we
already check that.