Ensure that pointer that is never used is not initialized to out of bounds value.
Only when y > 0 the decoding logic will use the pline pointer.
Only when y > 1 the decoding logic will use the ppline pointer.

But both pointers are always initialized to current line - 1 * stride
and current line - 2 * stride respectively. For the first two lines
that means that these pointers are initialized to pointers located
before the image data. This works fine in practice, since those
pointers are nevers used (since y must be > 0 and > 1 respectively for
the pointers to be used).

But Coverity can not figure that out, so it reported Coverity issues:

 * 487674 for jbig2_decode_generic_template1_TPGDON() and
 * 487684 for jbig2_decode_generic_template2_TPGDON().

For some reason Coverity did not spot the exact same code in

 * jbig2_decode_generic_template0_TPGDON() and
 * jbig2_decode_generic_template3_TPGDON().

This commit adds tests to all of these functions so that the pointers are
initailized to NULL instead of undefined values, and adds redundant check (y is
already checked) to ensure that the pointers are never used when set to NULL.