Check uses of strtok() for NULL return
Bug #708762 "Null pointer dereference when parsing dsc comments in call to gs_strtok from dsc_read_line"

The DSC parser wasn't checking for a NULL return from gs_strtok() when
looking for a specific number of tokens.

Checking the code base turned up a couple of other potential cases. It
isn't clear that any of these can be encountered normally but its better
to be safe.