PDF interpreter - improve checking of Optional Content Usage dctionaries
No bug report, came up in conversation on Discord. Previously we were
not checking to see if the Catalog had a /AS (usage application dict)
array or not, we simply applied any /Usage if the OCG was visible.

We now check the Catalog OCProperties dictionary for the presence of
a /AS array, if present then we iterate through the array looking for
Event dictionaries. If the device is a View device (basically the
display device) then we look for an Event dictionary with a /Event of
/View, otherwise we look for a dictionary with /Evenet of /Print.

If we find one, then we look for the preence in the dictionary of an
OCGs array, and if present we check its contents to see if the OCG we
are processing is listed. If it is, we check the OCG for a /Usage
dictionary and apply it.

Otherwise we ignore Usage dictionaries.

See /tests_private/comparefiles/BFAX35P5.pdf and
/tests_private/pdf/sumatra/2066_-_ocg_not_printed.pdf