Okidata dot matrix printer driver - detect invalid configuration
Bug #708768 "Global-buffer-underread due to missing check when using okiibm output device"

As noted in the bug report, some resolutions (roughly 0-30 and 150-210
assuming rounding) are invalid, this is signalled with a '-1' in the
graphics_modes_9 table. But the selected mode is never actually tested
to see if it is valid....

Check the selected mode, return an error if it is invalid. As an extra,
check the Y resolution too and throw an error if it is neitehr 72 nor
144 as these are the only supported resolutions.