pdfwrite - use 'hash' of filename from XUID to identify fonts.
Bug #708767 "missing glyphs after concatenating pdf files"

We msut have been using the filename differently to create the XUID in
the old PDF interpreter, as pdfwrite was not looking at the xvalues[1]
part of the XUID when determining font uniqueness.

The PDF interpreter says it is mimicking the old code, but clearly we
weren't using it that way.

Make pdfwrite consider the xvalues[1] (filename hash) when deciding if
two fonts are identicl. THis is in addition to the PDF obejct number
in xvalues[2].