pdfwrite - improve the XUID duplicate font detection
Chris pointed out that if the original font has an XUID then we extend
the XUID and insert our own at the start.

Cater for this by permitting the XUID to exceed 3 xvalues, but only
check and store the first three. We only use this for identifying
duplicate fonts, we don't embed the stored values.