transparency compositor - correctly dereference device
Bug #708783 "OOB read when using the display output device with a crafted PDF file"

There's no 'crafted PDF file' needed here, any PDF with transparency
will hit the problem, provided you are using the Linux display device
(NOT the x11alpha device).

pdev is a gx_device **, not a gx_device *, so it is necessary to
dereference it before passing to the function which expects a
gx_device *

I suspect the cast here may have been added by the original author to
(incorrectly) get rid of a compiler warning.

While passing I noted a compiler warning for an unused variable; get
rid of that too.